Started sketch of the queuing system
I started working on the inventory system, as well as the updating
queue. Basically, the updating queue is going to work as following: you
queue a Tickable for update (which is an implementation every object
using ticks should use), and then you add it to a UpdateQueue. You can
then update the queued in the order you have queued them in. This is to
prevent unnecessary updates of objects.